added comments to the file handler
